Split ScrolledEntry Backbone model test
REDMINE-20913
