+++ entry_types/scrolled/package/spec/editor/models/ScrolledEntry/supportsSectionWidths-spec.js
@@ -1,0 +1,36 @@
+import {ScrolledEntry} from 'editor/models/ScrolledEntry';
+import {factories} from 'pageflow/testHelpers';
+import {normalizeSeed} from 'support';
+
+describe('ScrolledEntry', () => {
+  describe('supportsSectionWidths', () => {
+    it('returns false by default', () => {
+      const entry = factories.entry(
+        ScrolledEntry,
+        {},
+        {entryTypeSeed: normalizeSeed()}
+      );
+
+      expect(entry.supportsSectionWidths()).toEqual(false);
+    });
+
+    it('returns true if theme has narrow section properties', () => {
+      const entry = factories.entry(
+        ScrolledEntry,
+        {},
+        {
+          entryTypeSeed: normalizeSeed({
+            themeOptions: {
+              properties: {
+                root: {
+                  narrowSectionTwoColumnInlineContentMaxWidth: '400px'
+                }
+              }
+            }
+          })
+      });
+
+      expect(entry.supportsSectionWidths()).toEqual(true);
+    });
+  });
+});
